type-aware `importX` extracted from base ESLint config & appended to `typeScript`
